Move get_unique_expression_symbols_as_strings to a new expression utilities module
